% Generated by roxygen2 (4.0.1): do not edit by hand
\name{angles}
\alias{angles}
\alias{as.angle}
\alias{as.bearing}
\alias{as.trig}
\title{Set or convert angles between conventions}
\usage{
as.bearing(x)

as.trig(x)

as.angle(x)
}
\arguments{
  \item{x}{vector of angles; either as numbers, in which
  case they are assumed to follow the appropriate
  convention; or of class circular, in which case the
  angles will be converted from their current convention to
  the convention determined by the function}
}
\description{
Set or convert angles between conventions
}
\details{
Bearings are angles from North, always positive, measured
clockwise, in degrees. Trigonometric angles are angles from
the horizontal, measured counter-clockwise, in radians.
"Angles" are trigonometric angles but in degrees.
}
\seealso{
\code{\link{circular}} and
\code{\link{conversion.circular}} in package
\code{circular}
}

