% Generated by roxygen2 (4.0.1): do not edit by hand
\name{assemble}
\alias{assemble}
\title{Assemble data of a given kind from several deployments}
\usage{
assemble(pattern, ids = NULL, deploy.dir = NULL)
}
\arguments{
  \item{pattern}{pattern to look for in the file name;
  passed to \code{\link[base]{list.files}}}

  \item{ids}{deployment ids to limit the search to; if NULL
  (the default) get data from all deployments}

  \item{deploy.dir}{path to the directory containing
  deployments (and associated options). By default
  "deployments" in the current directory.}
}
\description{
Assemble data of a given kind from several deployments
}

