% Generated by roxygen2 (4.0.0): do not edit by hand
\name{circ.stats}
\alias{circ.stats}
\title{Descriptive statistics and Rayleigh test}
\usage{
circ.stats(x)
}
\arguments{
  \item{x}{vector of angles, of class \code{circular}}
}
\value{
A data.frame with columns \itemize{ \item{n}{sample size}
\item{mean}{mean angle} \item{variance}{angular variance}
\item{r}{Rayleigh r} \item{p.value}{p-value of Rayleigh's
test} }
}
\description{
Descriptive statistics and Rayleigh test
}

