% Generated by roxygen2 (4.0.0): do not edit by hand
\name{circular_dotplot}
\alias{circular_dotplot}
\title{Circular dotplot}
\usage{
circular_dotplot(x, bin = 5, ...)
}
\arguments{
  \item{x}{vector of angles (potentially of class
  circular). Assumed to be in degrees and within [0,360]}

  \item{bin}{bin width in degrees}

  \item{...}{passed to \code{geom_point}}
}
\description{
Circular dotplot
}

