% Generated by roxygen2 (4.0.2): do not edit by hand
\name{disc_assemble}
\alias{dassemble}
\alias{disc_assemble}
\title{Assemble data of a given kind from several deployments}
\usage{
disc_assemble(pattern, ids = NULL, deploy.dir = NULL)

dassemble(pattern, ids = NULL, deploy.dir = NULL)
}
\arguments{
  \item{pattern}{pattern to look for in the file name;
  passed to \code{\link[base]{list.files}}. Only
  \code{.csv} files are searched for.}

  \item{ids}{deployment ids to limit the search to; if NULL
  (the default) get data from all deployments}

  \item{deploy.dir}{path to the directory containing
  deployments (and associated options). By default
  "deployments" in the current directory.}
}
\description{
Assemble data of a given kind from several deployments
}
\examples{
# get example deployments included with the package
deploys <- system.file("extdata", "deployments", package = "discr")
disc_status(deploys)

# get all larvae tracks
head(disc_assemble("rotated_larvae_tracks", deploy.dir=deploys))

# get all GPS tracks
head(dassemble("gps", deploy.dir=deploys))
tail(dassemble("gps", deploy.dir=deploys))
}

