% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disc_camera_compass_angle}
\alias{disc_camera_compass_angle}
\title{Measure angle between camera and compass}
\usage{
disc_camera_compass_angle(dir, verbose = FALSE, ...)
}
\arguments{
  \item{dir}{deployment directory}

  \item{verbose}{output messages on the console when TRUE}

  \item{...}{passthrough argument}
}
\description{
Measure angle of analog compasses on images, use that as a
reference, compare it to the digital compass record and
deduce the angle between the camera and the digital
compass. This is mandatory to use the digital compass to
correct the tracks.
}

