% Generated by roxygen2 (4.0.1): do not edit by hand
\name{disc_check}
\alias{disc_check}
\title{Check discuss running environment}
\usage{
disc_check()
}
\description{
Checks wether the working directory is set, what the
options are, and wether the required executables are
installed. If some are not available it gives platform
specific tips on how to install them. The function has no
arguments.
}
\seealso{
\code{\link{disc_getwd}} to check the working directory,
\code{\link{disc_setwd}} to set the working directory,
\code{\link{disc_conf}} to read or set the options
}

