% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disc_conf}
\alias{dconf}
\alias{disc_conf}
\title{Project-level settings}
\usage{
disc_conf(..., file = "disc.conf", verbose = FALSE)

dconf(..., file = "disc.conf", verbose = FALSE)
}
\arguments{
  \item{...}{named arguments with settings to set}

  \item{file}{name of the configuration file, in the
  project directory}

  \item{verbose}{when TRUE, options are shown after being
  read}
}
\description{
Read and write project-level settings
}
\details{
Settings are read, in order, from the defaults, then the
project-level configuration file and finally from the
arguments provided in \code{...}. Then those settings are
set with the function \code{\link{options}} to be
accessible from any other function in discuss. Finally, the
non-default settings are stored in \code{file} in the
project directory.
}

