% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_extract.R
\name{disc_extract}
\alias{disc_extract}
\alias{disc_extract.default}
\alias{disc_extract.gopro}
\title{Raw data extraction}
\usage{
disc_extract(data, start, stop, dir, ...)

\method{disc_extract}{default}(data, start, stop, dir, ...)

\method{disc_extract}{gopro}(data, start, stop, dir, width = 1600,
  gray = FALSE, ...)
}
\arguments{
\item{data}{data (or metadata) for this sensor, read by \code{\link{disc_read}}.}

\item{start, stop}{start and stop times, as \code{\link[base]{POSIXct}} objects.}

\item{dir}{destination directory, where the extracted data should be stored.}

\item{...}{passed to the various methods.}

\item{width}{width to resize the images to, in pixels. When NULL, images are not resized.}

\item{gray}{logical; whether to convert the images to grayscale.}
}
\description{
Extract data corresponding to one deployment from the raw data recorded by a sensor over a leg.
}
\details{
All methods can rely on \code{data} having a \code{dateTime} column, of class \code{\link[base]{POSIXct}}. Based on this column, they should extract the data between \code{start} and \code{stop} and store it in the destination directory.
}
\seealso{
\code{\link{disc_read}} and \code{\link{disc_extract_deployments}}
}

