% Generated by roxygen2 (4.0.1): do not edit by hand
\name{disc_getwd}
\alias{dgetwd}
\alias{disc_getwd}
\title{Get discuss working directory}
\usage{
disc_getwd(warn = FALSE)

dgetwd(warn = FALSE)
}
\arguments{
  \item{warn}{when TRUE, emit a warning when the working
  directory is not set}
}
\description{
Get discuss working directory
}

