% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disc_message}
\alias{disc_message}
\alias{dmess}
\title{Write a progress message on the console}
\usage{
disc_message(...)

dmess(...)
}
\arguments{
  \item{...}{message components passed to \code{cat}}
}
\description{
This is similar to \code{message} except it will add nice
indentation and won't be bold
}
\keyword{internal}

