% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disc_read}
\alias{disc_read}
\alias{disc_read.cc}
\alias{disc_read.dst}
\alias{disc_read.ez}
\alias{disc_read.gt31}
\alias{disc_read.hobo}
\alias{disc_read.igotu}
\alias{disc_read.opentag}
\alias{disc_read.remora}
\alias{disc_read.trackstick}
\title{Raw data read}
\usage{
disc_read(file, ...)

\method{disc_read}{gt31}(file, ...)

\method{disc_read}{igotu}(file, ...)

\method{disc_read}{trackstick}(file, ...)

\method{disc_read}{dst}(file, ...)

\method{disc_read}{opentag}(file, ...)

\method{disc_read}{ez}(file, ...)

\method{disc_read}{remora}(file, ...)

\method{disc_read}{opentag}(file, ...)

\method{disc_read}{cc}(file, ...)

\method{disc_read}{hobo}(file, ...)
}
\arguments{
  \item{file}{path to the file to read}

  \item{...}{passed to \code{\link{read.table}} which does
  the actual reading}
}
\value{
A data.frame with a columns \code{dateTime} and other
columns that depend on the sensor
}
\description{
Read data from various sensors output somewhat uniformised
records
}

