% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disc_setwd}
\alias{disc_setwd}
\alias{dsetwd}
\title{Set discuss working directory}
\usage{
disc_setwd(dir, persistent = FALSE)

dsetwd(dir, persistent = FALSE)
}
\arguments{
  \item{dir}{character string giving the name of the
  directory to set as the working directory}

  \item{persistent}{boolean; wether to store that working
  directory as default (storing \code{options(disc.wd=...)}
  in \code{.Rprofile})}
}
\value{
Returns the working directory but invisibly. Sets the
working directory for the current session and for all
future sessions if \code{persistent=TRUE}.
}
\description{
The working directory is the directory where deployments
are stored. Data will be read from and written to this
directory. Project level settings will also be stored
there.
}

