% Generated by roxygen2 (4.0.1): do not edit by hand
\name{disc_split_deployments}
\alias{disc_split_deployments}
\title{Get data for deployments}
\usage{
disc_split_deployments(raw, ids = NULL, acclimation.time = 5,
  observation.time = 15, width = 1600)
}
\arguments{
  \item{raw}{path to the directory where the raw data and
  the deployment and daily logs are.}

  \item{ids}{deployment identifiers to extract; if NULL
  (the default) get all deployments}

  \item{acclimation.time}{duration of the acclimation time
  in minutes.}

  \item{observation.time}{duration of the observation
  period (after acclimation) in minutes.}

  \item{width}{width to resize the images to, in pixels.
  When NULL, images are not resized.}
}
\description{
Read the deployment logs, read necessary data in the raw
data directory, split data into folders per deployment,
based on date and time
}

