% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/disc_start_project.R
\name{disc_start_project}
\alias{disc_start_project}
\title{Create a new project architecture}
\usage{
disc_start_project(path = ".", name = "DISC", ...)
}
\arguments{
\item{path}{path where the project is to be created. Current directory by default}

\item{name}{name of the project. 'DISC' by default}
}
\description{
Create a new project architecture
}

