% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disc_stats}
\alias{disc_stats}
\title{Compute statistics (circular and activity-level) on a deployment's trajectory data and create plots}
\usage{
disc_stats(dir, bin.angle = 0, sub = 0, verbose = FALSE, ...)
}
\arguments{
  \item{dir}{path the to the deployment directory}

  \item{bin.angle}{precision at which to bin angles}

  \item{sub}{subsample positions every sub seconds to make
  them independent}

  \item{verbose}{output messages on the console when TRUE}

  \item{...}{passthrough argument}
}
\description{
Compute statistics (circular and activity-level) on a
deployment's trajectory data and create plots
}

