% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disc_track}
\alias{disc_track}
\title{Manually track a larva}
\usage{
disc_track(dir, sub = NULL, verbose = FALSE, ...)
}
\arguments{
  \item{dir}{deployment directory}

  \item{sub}{subsampling interval, in s}

  \item{verbose}{output messages on the console when TRUE}
}
\description{
Use the manual tracking plugin to record a track
}

