% Generated by roxygen2 (4.0.1): do not edit by hand
\name{disc_track_compass}
\alias{disc_track_compass}
\title{Measure angle of analog compasses on images}
\usage{
disc_track_compass(dir, sub = NULL, verbose = FALSE, ...)
}
\arguments{
  \item{dir}{deployment directory}

  \item{sub}{subsampling interval, in s}

  \item{verbose}{output messages on the console when TRUE}

  \item{...}{passthrough argument}
}
\description{
Measure angle of analog compasses on images
}

