% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getdd}
\alias{getdd}
\alias{setdd}
\title{Get or set discuss deployments directory}
\usage{
getdd(warn = FALSE)

setdd(dir = "deployments")
}
\arguments{
  \item{dir}{character string giving the path to the
  directory to set as the working directory; either
  relative or absolute. By default: "deployments"}
}
\value{
\code{getdd} returns the deployments directory;
\code{setdd} and sets the deployments directory for the
current session and returns it invisibly
}
\description{
This is the directory where deployments are stored (usually
as a set of numbered folders). Data will be read from and
written to this directory. Settings will also be stored
there.
}
\details{
If the standard behaviour of having a "deployments"
directory in the current working directory is not
satisfactory, \code{setdd} can be used to set it to
something else (an absolute path on another drive, another
relative path such as "data/deployments", etc.). To make
this choice persistent across sessions, set the option
\code{disc.deployments} in your \code{.Rprofile}.
}
\keyword{internal}

