% Generated by roxygen2 (4.0.0): do not edit by hand
\name{image_interval}
\alias{image_interval}
\title{Compute the average interval in full seconds between provided images}
\usage{
image_interval(img)
}
\arguments{
  \item{img}{vector of full paths to images}
}
\description{
Compute the average interval in full seconds between
provided images
}
\keyword{internal}

