% Generated by roxygen2 (4.0.0): do not edit by hand
\name{image_time}
\alias{image_time}
\title{Extract time from EXIF data}
\usage{
image_time(img, tz = "GMT")
}
\arguments{
  \item{img}{full path to one or several image files}

  \item{tz}{force a time zone (most time computations are
  relative so getting the time zone right probably does not
  matter)}
}
\description{
Extract time from EXIF data
}
\keyword{internal}

