\name{make_path}
\alias{make_path}
\title{Create a valid path from one or several path elements}
\usage{
make_path(...)
}
\arguments{
  \item{...}{elements to be coerced as character strings
  and pasted together to make a path}
}
\description{
Create a valid path from one or several path elements
}
\keyword{internal}

