% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/angles.R
\name{pol2car}
\alias{pol2car}
\title{Convert from polar to cardinal coordinates}
\usage{
pol2car(x, orig = c(0, 0))
}
\arguments{
  \item{x}{matrix or data frame with columns [theta,rho],
  with theta of class circular or in trigonometric reference}
  \item{orig}{vector with the x,y coordinates of the origin,
  defaults to (0,0)}
}
\value{
a data.frame with columns [x,y] in the same unit as rho
}
\description{
Convert from polar to cardinal coordinates
}
\examples{
pol2car(data.frame(theta=pi/2, rho=1))
pol2car(data.frame(pi, 1))
pol2car(data.frame(as.bearing(90), 1))
}
\seealso{
\code{link{car2pol}} for the reverse
}

