% Generated by roxygen2 (4.0.0): do not edit by hand
\name{read_compass}
\alias{read_compass}
\alias{read_compass.cc}
\alias{read_compass.ez}
\alias{read_compass.opentag}
\alias{read_compass.remora}
\title{Compass reading}
\usage{
read_compass(file, ...)

\method{read_compass}{ez}(file, ...)

\method{read_compass}{remora}(file, ...)

\method{read_compass}{opentag}(file, ...)

\method{read_compass}{cc}(file, ...)
}
\arguments{
  \item{file}{path to the file to read}

  \item{...}{passed to \code{\link{read.table}} which does
  the actual reading}
}
\value{
A data.frame with columns: dateTime, heading, and other
columns that depend on the compass model
}
\description{
Read data from various digital compass models and output a
somewhat uniformised gps log
}

