% Generated by roxygen2 (4.0.0): do not edit by hand
\name{read_ctd}
\alias{read_ctd}
\alias{read_ctd.dst}
\alias{read_ctd.opentag}
\title{CTD reading}
\usage{
read_ctd(file, ...)

\method{read_ctd}{dst}(file, ...)

\method{read_ctd}{opentag}(file, ...)
}
\arguments{
  \item{file}{path to the file to read}

  \item{...}{passed to \code{\link{read.table}} which does
  the actual reading}
}
\value{
A data.frame with columns: dateTime, heading, and other
columns that depend on the ctd model
}
\description{
Read data from various CTD models and output a somewhat
uniformised log
}

