% Generated by roxygen2 (4.0.0): do not edit by hand
\name{read_gps}
\alias{read_gps}
\alias{read_gps.gt31}
\alias{read_gps.igotu}
\alias{read_gps.trackstick}
\title{GPS reading}
\usage{
read_gps(file, ...)

\method{read_gps}{gt31}(file, ...)

\method{read_gps}{igotu}(file, ...)

\method{read_gps}{trackstick}(file, ...)
}
\arguments{
  \item{file}{path to the file to read}

  \item{...}{passed to \code{\link{read.table}} which does
  the actual reading}
}
\value{
A data.frame with columns: dateTime, lon, lat, and other
columns that depend on the GPS model
}
\description{
Read data from various GPS model and output a somewhat
uniformised gps log
}

