% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.circular.R
\name{summary.circular}
\alias{summary.circular}
\title{Descriptive statistics and Rayleigh test}
\usage{
\method{summary}{circular}(object, ...)
}
\arguments{
\item{object}{vector of angles, of class \code{circular}}

\item{...}{passthrough argument}
}
\value{
A data.frame with columns
\describe{
\item{n: }{sample size}
\item{mean: }{mean angle}
\item{se.mean: }{standard error of the mean angle}
\item{kappa: }{kappa parameter (concentration) of a Von Mises distribution fitted to the data}
\item{se.kappa: }{standard error of the kappa parameter}
\item{variance: }{angular variance (i.e. 1-r)}
\item{r: }{Rayleigh's r, quantifying concentration}
\item{p.value: }{p-value of the Rayleigh test}
}
}
\description{
Descriptive statistics and Rayleigh test
}
\seealso{
Functions in the \code{circular} package: \code{link[circular]{mean.circular}} and \code{link[circular]{rayleigh.test}}
}

