% Generated by roxygen2 (4.0.0): do not edit by hand
\name{assemble}
\alias{assemble}
\title{Assemble data of a given kind from several deployments}
\usage{
assemble(pattern, ids = NULL, verbose = FALSE)
}
\arguments{
  \item{pattern}{pattern to look for in the file name;
  passed to list.files}

  \item{ids}{deployment ids to limit the search to; if NULL
  (the default) get data from all deployments}
}
\description{
Assemble data of a given kind from several deployments
}

