% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/angles.R
\name{car2pol}
\alias{car2pol}
\title{Convert from cardinal to polar coordinates}
\usage{
car2pol(x, orig = c(0, 0))
}
\arguments{
  \item{x}{matrix or data frame with columns [x,y]}
  \item{orig}{vector with the x,y coordinates of the origin,
  defaults to (0,0)}
}
\value{
a data.frame with columns [theta,rho], with theta following
trigonometric conventions and rho in the same unit as the
input x and y
}
\description{
Convert from cardinal to polar coordinates
}
\examples{
car2pol(data.frame(x=1, y=1))
car2pol(data.frame(1, 0))
car2pol(data.frame(-1, 0))
}
\seealso{
\code{link{pol2car}} for the reverse
}

