% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disc}
\alias{disc}
\title{Run a disc analysis action}
\usage{
disc(ids, actions = c("calibrate", "track", "compass", "stats"), ...)
}
\arguments{
  \item{ids}{vector of deployments ids (correspond to
  folders in the working directory). Usually deployment ids
  are integers, which makes it easy to give ranges of
  deployments such as 1:10. But deployment ids can be
  anything.}

  \item{actions}{vector of actions to perform. Action names
  can be abbreviated}

  \item{...}{passed to the various actions functions}
}
\description{
Run a disc analysis action
}
\seealso{
\code{\link{disc_setwd}} to set the working directory
}

