% Generated by roxygen2 (4.0.2): do not edit by hand
\name{disc}
\alias{disc}
\title{Run a disc analysis action}
\usage{
disc(ids = NULL, actions = c("calibrate", "track", "correct", "stats"),
  deploy.dir = NULL, ...)
}
\arguments{
  \item{ids}{vector of deployments ids (correspond to
  folders in the working directory). Usually deployment ids
  are integers, which makes it easy to give ranges of
  deployments such as 1:10. But deployment ids can be
  anything.}

  \item{actions}{vector of actions to perform. Action names
  can be abbreviated. See below which actions are
  available.}

  \item{deploy.dir}{path to the directory containing
  deployments (and associated options). By default
  "deployments" in the current directory.}

  \item{...}{passed to the various actions functions. To
  see which arguments are available, see the help of the
  various action functions}
}
\description{
Run a disc analysis action
}
\details{
\describe{ \item{"camera compass angle"}{estimate the angle
between the top of the camera and the physical direction in
which the digital compass is set up in its housing by
detecting the angle of the analog compasses on a few frames
and comparing that with the record of the digital compass.
Is usually done on short, calibration-type deployments.}
\item{"calibrate"}{measure the position of the arena on the
first image of a deployment, to calibrate the trajectory
and convert distances from pixels on frame to mm. See the
"\code{disc.diameter}" option set by
\code{\link{disc_conf}}.} \item{"track"}{manually track the
larva(e) on each frame} \item{"compass"}{manually detect
the analog compass bearing when the digital compass does
not work} \item{"correct"}{correct larvae tracks according
to compass readings, to put them back in cardinal space}
\item{"stats"}{compute statistics on the positions of
larvae in the arena} }
}
\examples{
# get example deployments included with the package
deploys <- system.file("extdata", "deployments", package = "discr")
# copy them to a writable, temporary directory
temp <- tempdir()
file.copy(deploys, temp, recursive=TRUE)
dd <- paste0(temp, "/deployments/")

disc_status(dd)

\donttest{# perform some actions on deployment 1
disc(ids=1, actions=c("calibrate", "track"), sub=10, deploy.dir=dd)
# action names can be abbreviated
disc(1, actions=c("cor", "st"), deploy.dir=dd)
# several deployments can be handled successively
disc(1:2, actions=c("cor", "st"), deploy.dir=dd)
# arguments can be passed to action functions
read.csv(paste0(dd, "/1/stats.csv"))
disc(1, actions="stats", sub=10, deploy.dir=dd)
read.csv(paste0(dd, "/1/stats.csv"))
}
}
\seealso{
Other action.functions: \code{\link{disc_calibrate}};
\code{\link{disc_camera_compass_angle}};
\code{\link{disc_correct}}; \code{\link{disc_stats}};
\code{\link{disc_track_compass}}; \code{\link{disc_track}}
}

