% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disc_calibrate}
\alias{disc_calibrate}
\title{Calibrate aquarium position}
\usage{
disc_calibrate(dir, ...)
}
\arguments{
  \item{dir}{path to the deployment directory}

  \item{...}{passthrough argument}
}
\description{
Calibrate aquarium position
}

