% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disc_calibrate}
\alias{dcalib}
\alias{dcalibrate}
\alias{disc_calibrate}
\title{Calibrate aquarium position}
\usage{
disc_calibrate(id, ...)

dcalibrate(id, ...)

dcalib(id, ...)
}
\arguments{
  \item{id}{deployment identifier, i.e. the name of the
  directory where data is stored}
}
\description{
Calibrate aquarium position
}

