\name{disc_calibrate}
\alias{dcalib}
\alias{dcalibrate}
\alias{disc_calibrate}
\title{Calibrate aquarium position}
\usage{
disc_calibrate(dir, ...)

dcalibrate(dir, ...)

dcalib(dir, ...)
}
\arguments{
  \item{dir}{path the to the deployment directory}
}
\description{
Calibrate aquarium position
}

