% Generated by roxygen2 (4.0.1): do not edit by hand
\name{disc_calibrate}
\alias{disc_calibrate}
\title{Calibrate aquarium position}
\usage{
disc_calibrate(dir, verbose = FALSE, ...)
}
\arguments{
  \item{dir}{path to the deployment directory}

  \item{verbose}{output messages on the console when TRUE}

  \item{...}{passthrough argument}
}
\description{
Calibrate aquarium position
}

