% Generated by roxygen2 (4.0.1): do not edit by hand
\name{disc_calibrate}
\alias{disc_calibrate}
\title{Calibrate aquarium position}
\usage{
disc_calibrate(dir, verbose = FALSE, ...)
}
\arguments{
  \item{dir}{path to the deployment directory}

  \item{verbose}{output messages on the console when TRUE}

  \item{...}{passthrough argument}
}
\description{
Calibrate aquarium position
}
\examples{
# get example deployments included with the package
deploys <- system.file("extdata", "deployments", package = "discusr")
# copy them to a writable, temporary directory
temp <- tempdir()
file.copy(deploys, temp, recursive=TRUE)
dd <- paste0(temp, "/deployments/")
deploy1 <- paste0(dd, "1")  # digital compass

# run the action
disc_conf(deploy.dir=dd)
\donttest{disc_calibrate(dir=deploy1, verbose=TRUE)}
# inspect results
list.files(deploy1)
# note that the aquarium coordinates are remembered
disc_conf(deploy.dir=dd)
}
\seealso{
Other action.functions:
\code{\link{disc_camera_compass_angle}};
\code{\link{disc_correct}}; \code{\link{disc_stats}};
\code{\link{disc_track_compass}}; \code{\link{disc_track}};
\code{\link{disc}}
}

