% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disc_camera_compass_angle}
\alias{disc_camera_compass_angle}
\title{Measure the angle between the camera and the compass}
\usage{
disc_camera_compass_angle(dir, ...)
}
\arguments{
  \item{dir}{deployment directory}
}
\description{
Measure the angle between the camera and the compass
}

