% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/disc_camera_compass_angle.R
\name{disc_camera_compass_angle}
\alias{disc_camera_compass_angle}
\title{Measure angle between camera and compass}
\usage{
disc_camera_compass_angle(dir, sub = NULL, verbose = FALSE, ...)
}
\arguments{
  \item{dir}{deployment directory} \item{sub}{subsampling
  interval, in s} \item{verbose}{output messages on the
  console when TRUE} \item{...}{passthrough argument}
}
\description{
Measure angle of analog compasses on images, use that as
a reference, compare it to the digital compass record and
deduce the angle between the camera and the digital compass.
This is mandatory to use the digital compass to correct the
tracks.
}
\examples{
# get example deployments included with the package
deploys <- system.file("extdata", "deployments", package = "discr")
# copy them to a writable, temporary directory
temp <- tempdir()
file.copy(deploys, temp, recursive=TRUE)
dd <- paste0(temp, "/deployments/")
deploy1 <- paste0(dd, "1")

# run the action
disc_conf(deploy.dir=dd)
\donttest{disc_camera_compass_angle(dir=deploy1, sub=10, verbose=TRUE)}
# the angle is also stored in a file "angle_camera_compass.txt"
list.files(deploy1)
}
\seealso{
Other action.functions: \code{\link{disc_calibrate}};
\code{\link{disc_correct}}; \code{\link{disc_stats}};
\code{\link{disc_track_compass}}; \code{\link{disc_track}};
\code{\link{disc}}
}

