% Generated by roxygen2 (4.0.1): do not edit by hand
\name{disc_conf}
\alias{dconf}
\alias{disc_conf}
\alias{print.disc_conf}
\title{Project-level settings}
\usage{
disc_conf(..., deploy.dir = NULL, verbose = FALSE)

dconf(..., deploy.dir = NULL, verbose = FALSE)

\method{print}{disc_conf}(x, ...)
}
\arguments{
  \item{...}{named arguments with settings to set}

  \item{verbose}{when TRUE, options are shown after being
  read}

  \item{deploy.dir}{path to the directory containing
  deployments (and associated options). By default
  "deployments" in the current directory.}
}
\description{
Read and write project-level settings
}
\details{
Settings are read, in order, from the defaults, then the
configuration file in the deployments directory and finally
from the arguments provided in \code{...}. Then those
settings are set with the function
\code{\link[base]{options}} to be accessible from any other
function in \code{discusr}. Finally, the non-default
settings are stored in the file \code{"disc.conf"} in the
deployments directory.
}
\examples{
disc_conf()
# increase java memory to 2G
disc_conf(java_memory=2048)
# change diameter of arena to 40 cm (large DISC)
disc_conf(diameter=40)
}
\keyword{internal}

