% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disc_correct}
\alias{dcorrect}
\alias{disc_correct}
\title{Correct tracks (camera frame of reference) based on rotation of DISC to get polar-relevant coordinates}
\usage{
disc_correct(dir, ...)

dcorrect(dir, ...)
}
\arguments{
  \item{dir}{path the to the deployment directory}
}
\description{
Correct tracks (camera frame of reference) based on
rotation of DISC to get polar-relevant coordinates
}

