% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disc_correct}
\alias{disc_correct}
\title{Correct tracks (camera frame of reference) based on rotation of DISC to get polar-relevant coordinates}
\usage{
disc_correct(dir, verbose = FALSE, ...)
}
\arguments{
  \item{dir}{path the to the deployment directory}

  \item{verbose}{output messages on the console when TRUE}
}
\description{
Correct tracks (camera frame of reference) based on
rotation of DISC to get polar-relevant coordinates
}

