% Generated by roxygen2 (4.0.2): do not edit by hand
\name{disc_correct}
\alias{disc_correct}
\title{Correct tracks (camera frame of reference) based on rotation of DISC to get polar-relevant coordinates}
\usage{
disc_correct(dir, camera.compass.angle = NULL, verbose = FALSE, ...)
}
\arguments{
  \item{dir}{path the to the deployment directory}

  \item{camera.compass.angle}{angle between the top of the
  camera and the N of the digital compass. If NULL, should
  be read from a file named "angle_camera_compass.txt" in
  the deployment directory}

  \item{verbose}{output messages on the console when TRUE}

  \item{...}{passthrough argument}
}
\description{
Correct tracks (camera frame of reference) based on
rotation of DISC to get polar-relevant coordinates
}
\examples{
# get example deployments included with the package
deploys <- system.file("extdata", "deployments", package = "discr")
# copy them to a writable, temporary directory
temp <- tempdir()
file.copy(deploys, temp, recursive=TRUE)
dd <- paste0(temp, "/deployments/")
deploy1 <- paste0(dd, "1")  # digital compass
deploy2 <- paste0(dd, "2")  # analog compass

# run the action
disc_conf(deploy.dir=dd)
disc_correct(dir=deploy1, verbose=TRUE)
# inspect results
list.files(deploy1)
head(read.csv(paste0(deploy1, "/larvae_tracks.csv")))
head(read.csv(paste0(deploy1, "/rotated_larvae_tracks.csv")))
tail(read.csv(paste0(deploy1, "/rotated_larvae_tracks.csv")))

disc_correct(dir=deploy2, verbose=TRUE)
list.files(deploy2)
}
\seealso{
Other action.functions: \code{\link{disc_calibrate}};
\code{\link{disc_camera_compass_angle}};
\code{\link{disc_stats}}; \code{\link{disc_track_compass}};
\code{\link{disc_track}}; \code{\link{disc}}
}

