% Generated by roxygen2 (4.0.1): do not edit by hand
\name{disc_read}
\alias{disc_read}
\alias{disc_read.cc}
\alias{disc_read.dst}
\alias{disc_read.ez}
\alias{disc_read.gopro}
\alias{disc_read.gt31}
\alias{disc_read.hobo}
\alias{disc_read.igotu}
\alias{disc_read.opentag}
\alias{disc_read.remora}
\alias{disc_read.trackstick}
\title{Raw data read}
\usage{
disc_read(dir, ...)

\method{disc_read}{gt31}(dir, ...)

\method{disc_read}{igotu}(dir, ...)

\method{disc_read}{trackstick}(dir, ...)

\method{disc_read}{dst}(dir, ...)

\method{disc_read}{opentag}(dir, ...)

\method{disc_read}{ez}(dir, ...)

\method{disc_read}{remora}(dir, ...)

\method{disc_read}{opentag}(dir, ...)

\method{disc_read}{cc}(dir, ...)

\method{disc_read}{hobo}(dir, ...)

\method{disc_read}{gopro}(dir, ...)
}
\arguments{
  \item{dir}{path to the directory in which the files to
  read are}

  \item{...}{passed to \code{\link{read.table}} which does
  the actual reading}
}
\value{
A data.frame with a columns \code{dateTime} and other
columns that depend on the sensor
}
\description{
Read data from various sensors output somewhat uniformised
records
}

