% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc_read.R
\name{disc_read}
\alias{disc_read}
\alias{disc_read.cc}
\alias{disc_read.compassOpentag}
\alias{disc_read.compassRemora}
\alias{disc_read.ctdOpentag}
\alias{disc_read.dst}
\alias{disc_read.dstctd}
\alias{disc_read.ez}
\alias{disc_read.gopro}
\alias{disc_read.goproVideo}
\alias{disc_read.gt31}
\alias{disc_read.gw52}
\alias{disc_read.hobo}
\alias{disc_read.igotu}
\alias{disc_read.remoraAr}
\alias{disc_read.trackstick}
\title{Raw data reading}
\usage{
disc_read(dir, ...)

\method{disc_read}{gt31}(dir, ...)

\method{disc_read}{igotu}(dir, ...)

\method{disc_read}{trackstick}(dir, ...)

\method{disc_read}{gw52}(dir, ...)

\method{disc_read}{dst}(dir, ...)

\method{disc_read}{dstctd}(dir, ...)

\method{disc_read}{ctdOpentag}(dir, ...)

\method{disc_read}{ez}(dir, ...)

\method{disc_read}{compassRemora}(dir, ...)

\method{disc_read}{compassOpentag}(dir, ...)

\method{disc_read}{cc}(dir, ...)

\method{disc_read}{hobo}(dir, ...)

\method{disc_read}{gopro}(dir, ...)

\method{disc_read}{goproVideo}(dir, ...)

\method{disc_read}{remoraAr}(dir, ...)
}
\arguments{
\item{dir}{path to the directory containing the files to read.}

\item{...}{passed to other functions that do the actual reading, most often \code{\link[utils]{read.table}}.}
}
\description{
Read data from various sensors and output somewhat uniformed records.
}
\details{
All methods of this generic function should takes the path to a directory (where the data is stored) as input and give a data.frame as output, with at least a column called \code{dateTime} of class \code{\link[base]{POSIXct}} containing the date and time of that record. This column will be used to shift the data by a user configured offset (in \code{\link{disc_extract_deployments}}) and synchronise sensors. The rest of the columns depend on the sensors. The only constraint is for compass-type data to have a column named \code{heading}.

For instruments that output binary data (videos, audio files, etc.), the \code{disc_read} method should provide time stamps and paths to the various files. This information is then used by the appropriate \code{disc_extract} method to actually move/split/etc. the leg-level binary file into a file for each deployment.
}
\seealso{
\code{\link{disc_extract}} and \code{\link{disc_extract_deployments}}
}

