% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/disc_read.R
\name{disc_read}
\alias{disc_read}
\alias{disc_read.cc}
\alias{disc_read.compass_opentag}
\alias{disc_read.ctd_opentag}
\alias{disc_read.dst}
\alias{disc_read.ez}
\alias{disc_read.gopro}
\alias{disc_read.gt31}
\alias{disc_read.hobo}
\alias{disc_read.igotu}
\alias{disc_read.remora}
\alias{disc_read.trackstick}
\title{Raw data reading}
\usage{
disc_read(dir, ...)

\method{disc_read}{gt31}(dir, ...)

\method{disc_read}{igotu}(dir, ...)

\method{disc_read}{trackstick}(dir, ...)

\method{disc_read}{dst}(dir, ...)

\method{disc_read}{ctd_opentag}(dir, ...)

\method{disc_read}{ez}(dir, ...)

\method{disc_read}{remora}(dir, ...)

\method{disc_read}{compass_opentag}(dir, ...)

\method{disc_read}{cc}(dir, ...)

\method{disc_read}{hobo}(dir, ...)

\method{disc_read}{gopro}(dir, ...)
}
\arguments{
  \item{dir}{path to the directory in which the files to read
  are} \item{...}{passed to \code{\link[utils]{read.table}}
  which does the actual reading}
}
\description{
Read data from various sensors output somewhat uniformed
records
}
\details{
\code{\link{disc_extract_deployments}} reads data
from all sensors defined in the leg log file. To do
so it looks for an appropriate method for the generic
function \code{disc_read}, i.e. a function named
\code{disc_read.nameofsensor}. Do read data from a new
sensor, one just need to define such as function, which
takes the path to a directory (where the data is stored)
as input and gives a data.frame as output, with at least a
column called "\code{dateTime}" holding the date and time in
the default format for \code{link[base]{as.POSIXct}} (i.e.
"YYYY-MM-DD HH:MM:SS"). The rest of the columns depend on
the sensors. The only constraint is for compass-type data to
have a column named "\code{heading}"
}

