% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disc_split_deployments}
\alias{disc_split_deployments}
\title{Get data for deployments}
\usage{
disc_split_deployments(raw, dest = disc_getwd(), ids = NULL,
  acclimation.time = 5, observation.time = 15)
}
\arguments{
  \item{raw}{path to the directory where the raw data and
  the deployment and daily logs are.}

  \item{dest}{path to destination directory, where
  deployment folders will be stored. Will be created if it
  does not exist. Is set to discuss working directory by
  default.}

  \item{ids}{deployment identifiers to extract; if NULL
  (the default) get all deployments}

  \item{acclimation.time}{duration of the acclimation time
  in minutes.}

  \item{observation.time}{duration of the observation
  period (after acclimation) in minutes.}
}
\description{
Read the deployment logs, read necessary data in the raw
data directory, split data into folders per deployment,
based on date and time
}
\details{
The csv file with the deployment log. Should have, at
least, columns named 'id', 'start_date', 'start_time',
'stop_date', 'stop_time'. Date should be formatted as
YYYY-MM-DD and time as HH:MM:SS. Start date and time are
when the DISC is beginning to being lowered in the water.
Stop date and time are when the DISC is back on the boat.
}

