% Generated by roxygen2 (4.0.0): do not edit by hand
\name{disc_stats}
\alias{disc_stats}
\alias{dstats}
\title{Perform Statistics (circular and activity-level) on a deployment's trajectory data, create plots}
\usage{
disc_stats(dir, ...)

dstats(dir, ...)
}
\arguments{
  \item{dir}{path the to the deployment directory}
}
\description{
Perform Statistics (circular and activity-level) on a
deployment's trajectory data, create plots
}

