% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/disc_stats.R
\name{disc_stats}
\alias{disc_stats}
\title{Compute statistics (circular and activity-level) on a deployment's trajectory data and create plots}
\usage{
disc_stats(dir, bin.angle = 0, sub = NULL, verbose = FALSE, ...)
}
\arguments{
  \item{dir}{path the to the deployment directory}
  \item{bin.angle}{precision at which to bin angles}
  \item{sub}{subsample positions every sub seconds to make
  them independent} \item{verbose}{output messages on the
  console when TRUE} \item{...}{passthrough argument}
}
\description{
Compute statistics (circular and activity-level) on a
deployment's trajectory data and create plots
}
\examples{
# get example deployments included with the package
deploys <- system.file("extdata", "deployments", package = "discr")
# copy them to a writable, temporary directory
temp <- tempdir()
file.copy(deploys, temp, recursive=TRUE)
dd <- paste0(temp, "/deployments/")
deploy1 <- paste0(dd, "1")

# run the action
disc_conf(deploy.dir=dd)
disc_stats(dir=deploy1, verbose=TRUE)
# inspect results
list.files(deploy1)
read.csv(paste0(deploy1, "/stats.csv"))

# subsample positions
disc_stats(dir=deploy1, sub=10, verbose=TRUE)
read.csv(paste0(deploy1, "/stats.csv"))
# Note the difference in n compared to above
}
\seealso{
Other action.functions: \code{\link{disc_calibrate}};
\code{\link{disc_camera_compass_angle}};
\code{\link{disc_correct}};
\code{\link{disc_track_compass}}; \code{\link{disc_track}};
\code{\link{disc}}
}

