% Generated by roxygen2 (4.0.1): do not edit by hand
\name{disc_status}
\alias{disc_status}
\alias{dstatus}
\title{Status of the working directory}
\usage{
disc_status(dir = NULL)

dstatus(dir = NULL)
}
\arguments{
  \item{dir}{path to the deployments directory. Assumed to
  be in the working directory by default.}
}
\description{
Give information about deployments (number of pictures,
metadata files, data files, etc.)
}
\seealso{
\code{\link{disc_setwd}} to set the working directory
option.
}

