% Generated by roxygen2 (4.0.2): do not edit by hand
\name{disc_status}
\alias{disc_status}
\alias{dstatus}
\title{Status of the deployments directory}
\usage{
disc_status(deploy.dir = NULL)

dstatus(deploy.dir = NULL)
}
\arguments{
  \item{deploy.dir}{path to the directory containing
  deployments (and associated options). By default
  "deployments" in the current directory.}
}
\description{
Give information about deployments (number of pictures,
metadata files, data files, etc.)
}
\examples{
# get example deployments included with the package
deploys <- system.file("extdata", "deployments", package = "discr")

disc_status(deploy.dir=deploys)
}

