% Generated by roxygen2 (4.0.1): do not edit by hand
\name{disc_track}
\alias{disc_track}
\title{Manually track a larva}
\usage{
disc_track(dir, sub = NULL, verbose = FALSE, ...)
}
\arguments{
  \item{dir}{deployment directory}

  \item{sub}{subsampling interval, in s}

  \item{verbose}{output messages on the console when TRUE}

  \item{...}{passthrough argument}
}
\description{
Use the manual tracking plugin to record a track
}
\seealso{
Other action.functions: \code{\link{disc_calibrate}};
\code{\link{disc_camera_compass_angle}};
\code{\link{disc_correct}}; \code{\link{disc_stats}};
\code{\link{disc_track_compass}}; \code{\link{disc}}
}

