% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/disc_track.R
\name{disc_track}
\alias{disc_track}
\title{Manually track a larva}
\usage{
disc_track(dir, sub = NULL, verbose = FALSE, ...)
}
\arguments{
\item{dir}{deployment directory}

\item{sub}{subsampling interval, in s}

\item{verbose}{output messages on the console when TRUE}

\item{...}{passthrough argument}
}
\description{
Use the manual tracking plugin to record a track
}
\examples{
# get example deployments included with the package
deploys <- system.file("extdata", "deployments", package = "discr")
# copy them to a writable, temporary directory
temp <- tempdir()
file.copy(deploys, temp, recursive=TRUE)
dd <- paste0(temp, "/deployments/")
deploy1 <- paste0(dd, "1")

# run the action
disc_conf(deploy.dir=dd)
\donttest{disc_track(dir=deploy1, verbose=TRUE)}
# inspect results
list.files(deploy1)
head(read.csv(paste0(deploy1, "/larvae_tracks.csv")))
}
\seealso{
Other action.functions: \code{\link{disc_calibrate}};
  \code{\link{disc_camera_compass_angle}};
  \code{\link{disc_correct}}; \code{\link{disc_stats}};
  \code{\link{disc_track_compass}}; \code{\link{disc}}
}

