% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.circular.R
\name{summary.circular}
\alias{summary.circular}
\title{Descriptive statistics and Rayleigh test}
\usage{
\method{summary}{circular}(object, ...)
}
\arguments{
\item{object}{vector of angles, of class \code{circular}}

\item{...}{passthrough argument}
}
\value{
A data.frame with columns
\describe{
\item{n: }{sample size}
\item{mean: }{mean angle}
\item{variance: }{angular variance}
\item{r: }{Rayleigh r}
\item{p.value: }{p-value of Rayleigh's test}
}
}
\description{
Descriptive statistics and Rayleigh test
}
\seealso{
Functions in the \code{circular} package: \code{link[circular]{mean.circular}} and \code{link[circular]{rayleigh.test}}
}

